// types/dbState.d.ts
export interface DBState {
  name: string;
  type: string;
  color: string;
  priority: string;
  restriction: string;
  a_rate: string;
  b_rate: string;
  c_rate: string;
  d_rate: string;
  e_rate: string;
  hold_turn: string;
  auto_release_prob: string;
  release_by_damage: string;
  affect_type: string;
  affect_attack: string;
  affect_defense: string;
  affect_spirit: string;
  affect_agility: string;
  reduce_hit_ratio: string;
  avoid_attacks: string;
  reflect_magic: string;
  cursed: string;
  battler_animation_id: string;
  restrict_skill: string;
  restrict_skill_level: string;
  restrict_magic: string;
  restrict_magic_level: string;
  hp_change_type: string;
  sp_change_type: string;
  message_actor: string;
  message_enemy: string;
  message_already: string;
  message_affected: string;
  message_recovery: string;
  hp_change_max: string;
  hp_change_val: string;
  hp_change_map_steps: string;
  hp_change_map_val: string;
  sp_change_max: string;
  sp_change_val: string;
  sp_change_map_steps: string;
  sp_change_map_val: string;
  easyrpg_immune_states: string;
}